Add ability to remove Views
If a View still has any Components that are exclusive to it (not also
visible in another View), then we do not want to allow removal of that
View to prevent orphaning Components, making them no longer visible in
any View. If there are View objects on the canvas that are not in any
other Views, this is fine, since Views are top-level objects of their
own (and all Views start out as "orphans" as soon as they are created).

This means that when we are applying a batch of changes in the rebaser,
we need to make sure that if Components were added to a View that the
batch of updates isn't handling the removal of, we want to skip the
removal of the View to prevent orphaning the "new" Components.