Add audit database migrations
This commit adds audit database migrations. They are explicitly not part
of the ServicesContext nor DalContext.

This commit also moves the audit logs stream contents to a new submodule
in the corresponding crate as well as deletes an unused NATS client
method in sdf-server.

We also use the DBNAME constant for layer db configuration in sdf now,
like we do with the audit database.

Co-authored-by: Fletcher Nichol <fletcher@systeminit.com>
Co-authored-by: John Watson <john@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>