fix(layer-cache): add size hinting when inserting Rust types into caches
This change addresses an issue when inserting a Rust type into a Foyer
cache, where the memory used by that rust variable was not being
accurately calculated. Prior to this change, the `size_of_val()` would
dereference one level of "smart pointers" (i.e. deref an `Arc`, or
`Vec`, or `String`), but no more levels. This led to an undercount of
memory for that Rust value. For example, perhaps a workspace snapshot
that was 30mb in size would be incorrectly calculated as 5mb.

As determining the size of an arbitrary Rust object is surprisingly
difficult, we opted to use the *uncompressed* size of serialized
Postcard bytes as our "size hint" (note that Postcard adds very little
extra information into the stream of bytes, so could be considered as a
very close approximation) and to provide this size hint to Foyer's
`with_weighter()` configuration. In our caches, when we have a
`MaybeDeserialized::RawBytes` variant, the size of its Bytes is the size
of that value (and so can be used directly).

Co-authored-by: Scott Prutton <scott@systeminit.com>
Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>