Read from the audit database
This commit ensures that we read from the audit database for the audit
logs interface rather than directly from the NATS JetStream stream. This
is a drop-in replacement that does not result in a change to the front
end. The user experience, however, should be vastly improved with one
caveat that we'll cover at the end.

To accomplish this, we always read from the database in SDF and in
integration tests, but enabling the audit logs app for forklift will
still be opt-in for production use cases. However, for CI integration
tests, local integration tests and development builds with buck2 and
cargo, forklift will have the audit logs app enabled by default.

This commit removes the ability to read from the stream in the dal
wholesale and also uses the new "AuditLogRow" struct and its methods
to read from the database. We rely on the compound index for the
"audit_logs" table in the audit database for quick querying.

Context:

This switchover was implemented for a variety of reasons, including
but not limited to the following:

- Data correctness in the frontend (e.g. the number of rows you ask for
  should now be what you get, or at least however many exist for your
  currently selected change set)
- Performance
- Rich query semantics
- Data segregation by database

Test related refactoring:

Ever wondered why our services often have a "from_config" method that
calls a "from_services" method? It's because we don't want services
bootstrapping their own duplicate inner services if we don't need to in
tests. Now, forklift is able to take in an active audit database context
that's shared with the test author's audit database context. Well, the
inner PgPool is shared anyway.

Known issue:

If you manipulate filters in the audit logs UI, specifically clicking
"Clear Filters", you will notice a significant slowdown. This is not
specific to this PR and is an issue in "app/web" that will be solved
in a follow-up PR.

Co-authored-by: Fletcher Nichol <fletcher@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>