feat: management functions understand views
Management funcs are now run in the context of a view. They can only
create components in that view. But they can manage components in other
views. The position of created components is relative to the manager
component, but the position specified in updates is absolute.

If a manager component is in more than one view, a the "play" button
will open a dropdown allowing you to select the view to run the function
in.

This also makes position (x,y) optional, allowing you to update just the
size of a component.