chore(forklift): add tuning for concurrent audit log database inserts
This change adds a tunable in forklift's configuration called
`insert_concurrency_limit` to control how many concurrent inserts are
running.

This can be set in a `forklift.toml` with:

```toml
[audit]
insert_concurrency_limit = 4
```

Or in an environment variable with:

```sh
SI_FORKLIFT__AUDIT__INSERT_CONCURRENCY_LIMIT=4
```

Longer term, each microservice that forklift runs should have
its own tunable for its own concurrency as one broad brush stroke won't
be sufficient in the various environments. Additionally the name for the
idea of a concurrency limit has `insert_` in it because currently we
render the exact same `config.toml` template in all environments and for
all services. This is one way for the moment to be very clear as to what
this tunes (as read from a more "global" context).

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>