feat: migrate local module cache on sdf boot, in dev mode
Use the same technique for "development run" detection used to set up
the signing keys to set a dev_mode flag on the SDF config. If this
dev_mode flag is true, and we are running migrations, also update the
local module cache. This ensures we have an asset list locally on SDF
boot.