Ensure filters do not affect audit table contents
This commit ensures that changing filters does not affect the contents
of the inner table used in the Audit Log dashboard. This does not fully
fix the performance issues observed when using filters in the Audit Log
dashboard, but does improve them.

Now, the data for the table is only changed when the total set of Audit
Log entries loaded changes. This commit takes advantage of TanStack's
built-in filtering features. However, this commit does not fully take
advantage of every TanStack feature at our disposal. There is likely
more room for improvement.

Additionally, there is a now a distinct message for scenarios where
there are no audit logs in the workspace (e.g. brand new workspace).

Signed-off-by: Nick Gerace <nick@systeminit.com>