feat: support multiple jwt public keys
Support two JWT keys. One used for signing (the primary private key),
and an extra one for verification, so that we can rotate keys without
downtime, and then remove the "secondary" if desired. It also now
possible to specify the key algorithm. ES256 and RS256 are supported but
RS256 will be removed.

Adds a new crate, `si-jwt-public-key` (since this logic needs to be
shared between sdf and the module-index).