Centralize IDs to "si-id"
This commit centralizes (mostly) all IDs in the codebase to a new crate
called "si-id". This work was continued from #4746 and original research
in a now-deleted branch, "nick/3b20c5e".

Methodology:

The new "id!" macro in "si-id" contains a superset of all methods and
types contained in the old macros. Yes, this means that the unsightly
"NONE" type is available for all ID types now. However, we've developed
good habits and will be able to splice out this usages in follow-up
work.

Potential future work:

- Clean-up ID call sites (e.g. reduce "pub use" statements)
- Migrate all IDs in "si-layer-cache" (were not reliant on a macro)
- Migrate all IDs in "module-index-server" (were not reliant on a macro)
- Remove all usages of the third party ulid crate and go through "si-id"
- Add another macro in "si-id" that contains the "NONE" type (the main
  macro should not have it)
- Add another macro in "si-id" that contains the postgres types (the
  main macro should not have it)

Signed-off-by: Nick Gerace <nick@systeminit.com>