fix: schemas should be created with lineage_id == id
On demand asset installation can produce identical Schema nodes by id,
but they will have different lineage ids. Since the lineage ids are
different, we will treat them as new nodes on the graph. The idempotency
of our updates usually means this is fine. But the interaction with the
code to ensure there is only one default edge caused an issue.

This fixes both, by ensuring new schemas are always created with the
a lineage id equal to the schema id. This will ensure we compare the
nodes instead of treating the schema node as a new node.

It also ensures that if a new edge update comes in that sets a schema
variant to the default, but that variant is already default, we preserve
that default.