feat: mgmt fns can create views, create components in views
The return type for management functions has been extended to support
creating views, and creating new components in views other than the
current view.

The "geometry" property on a create can now *either* be a plain
geometry object, which will create the component in the current view, or
it can be a map of view names to geometries, which will create the
component in the specified view with that geometry.

Additionally, you can add to your ops return an object like so:

```
ops: {
  views: {
    create: ["view a", "view b", "view c"]
  }
}
```

to create new views in a mangement function. The views will be available
for create and update operations in the management func itself.