Add ability to set approvers on views in sdf
This commit adds the ability to set approvers on the views with a new
set of sdf routes, backed by a new integration test among existing ones.

To support this change, we now divide ApprovalRequirements into two
separate categories: explicit and virtual.

- Explicit requirements are derived from ApprovalRequirementDefinition
  nodes (new to WorkspaceSnapshotGraphV4!) on the graph from any node
  with an outgoing HasApprovalRequirement edge
- Virtual requirements are derived from examining a list of changes on
  the fly and are useful for determining if meta changes to permissions
  require approvals themselves (e.g. if you add a
  ApprovalRequirementDefinition, we need to determine who can approve that
  change)

Both types of requirements share an inner "rule" type with covers
the universal elements of all approval requirements:

- The ID of the entity requiring an approval
- The kind of entity requiring an approval
- The minimum number of approvals for this rule
- The approvers of this rule, which can either be individual users or
  a "permission lookup" bag used to query SpiceDB for who can approve
  the requirement

Back to virtual requirements: not only are they useful for determining
meta changes to permissions, but also for creating requirements when
nodes are deleted and creating the checksum for new approvals involving
deleted nodes. In fact, that is why checksum calculatation now requires
the hashes alongside the IDs... you can't get the merkle tree hash for a
deleted node!

As a result of these changes, detecting changes between two graphs is
now fallible (we need to get the node weight in order to get the merkle
tree hash).

Other changes in this commit include, but are not limited to, the
following:

- Use UserPk instead of String for return type involving frontend
  approvals
- Re-order enum variants and their discriminants where possible when
  dealing with ApprovalRequirementDefinition and HasApprovalRequirement
- Rename "geometry" dummy data to "view" in deserialize graph
  integration test because it could be confusing (although it is
  entirely cosmetic for the purposes of the test)

Co-authored-by: Jacob Helwig <jacob@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>