build: alter Docker image multi-arch promotion for certain images
This change changes the default machine architectures that are expected
to have Docker images being built and published to Docker Hub for tag
promotion (in the `si/merge-main` CI pipeline).

By default, it's expected that an ARM and an X86-64 image will be
produced. In this change it's now possible to change that default
assumption and to only specify the machine architectures expected to
have been published. The underlying Python script and corresponding
`docker_image_promote` rule already had this capability but the
`docker_image` Buck2 macro wasn't able to handle this prior to this
change.

This change works in concert with changes to `si-ci` which will filter
out non-critical artifact builds in the `si/merge-queue` pipeline.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>