use si.error.message instead of error in spans
When we do something like `error=?err` the `error` attribute in Honeycomb is mapped to a bool, so we lose the context. By sending the error struct to `si.error.message` we get more context. This would have been helpful when troubleshooting this week, so I tried to update them everywhere for consistency sake.