Rename HasApprovalRequirement for clarity
This commit renames the HasApprovalRequirement EdgeWeight to
ApprovalRequirementDefinition for clarity. This rename removes hints
of directionality and specifically refers to the
ApprovalRequirementDefinition node weight.

This commit also fixes two debug output locations for the new node
weight as well as changes the order (with a comment) for the approver
enum variants.

Co-authored-by: Jacob Helwig <jacob@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>