Address component CRUD in views
This commit addresses component CRUD operations in views. Specifically,
if you create, update or delete a component with a view with approvals,
this commit contains the test necessary for ensuring this works.

In addition to these changes, the dal wrapper module has been split
into a sub-module for change set approvals. Within it, the calculator
struct has been replaced with a single function instead of its three
public functions that served the same purpose.

This commit also includes the beginning of explicit actions CRUD within
a view as well (starting with category node fetching).

Given the amount of concurrent work that has been happening, here is a
paper trail of all relevant PRs to the fine grained access control
feature thus far:
- #5330
- #5324
- #5279
- #5278
- #5144

Co-authored-by: Jacob Helwig <jacob@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>