Support rejection in fine grained access control
This commit adds the ability to reject a change set with the fine
grained access control feature enabled. Essentially, rejection is a
"social" feature in the sense that it does not affect what can be done
with a change set. However, in this world, we still have the ability
to determine if a rejection is "stale" and what requirements it is
applicable for.

How is this a one-liner change in the backend? Approval checksum
calculation is automatic and the overall creation flow is already set up
for different status types. In addition, when checking if requirements
have been satisfied, we already check if the status is "approved", even
though there was only one enum variant at the time that code was
written.

That all being said, this commit introduces the ability to reject in
the frontend and a corresponding integration test has been amended to
double check that this works.

One additional change: remove the response payload when applying a
change set in the change sets store because sdf does not return a
payload.

Signed-off-by: Nick Gerace <nick@systeminit.com>