Move list workspace users of out the admin routes
This commit moves the "list workspace users" functionality out of the
admin routes and into the workspace routes. We do not want to use the
admin DalContext nor the admin store unless we're in the AdminDashboard
and this change allows anyone in the workspace to list workspace users,
which is required for fine grained access control (and is generally
useful).

However, the auth store location in the frontend and the workspace
route addition in the backend are not necessarily the best places for
this functionality to live either. This commit serves as a quick
stepping stone that can be refactored when the domain begins to boil
over.

Additional work in this commit includes, but is not limited to cleaning
up workspace API error types and general re-organization of errors for
readability and clarity. In addition, a long-standing "TODO" to only
react to the approval status WsEvent if it matches our current change
set has been addressed here (which is feature flagged).

Signed-off-by: Nick Gerace <nick@systeminit.com>