Protect against unforced apply without reqs met
This commit adds a new route for applying a change set with fine
grained access control. This route contains new dal wrapper logic to
ensure that applying a change set (unforced) only happens if all
requirements have been met. There's a new integration test for this as
well.

Other changes include, but are not limited to, the following:
- Split apply prepare step into two versions because sdf now handles
  the protection layer
- Rename the "change_set_approval" module to "change_set" in dal wrapper
- Add sdf-test helpers for mimicking protected apply behavior in the
  route, which is used by the new test

Signed-off-by: Nick Gerace <nick@systeminit.com>