Execute the kill execution test by killing it
This commit just flat out removes the kill execution test. Honestly? I
try to avoid removing test coverage outright. It's like antithetical to
who I am. However, this... THIS... DAMN. TEST. It's causing unrelenting
pain and has for a sizable chunk of this repository's life. In a poetic
end to the kill execution test, we are executing it via guillotine
(metaphorical). I'd say we don't deserve this test, but frankly, I've
concluded that it doesn't deserve us.

Signed-off-by: Nick Gerace <nick@systeminit.com>