feat: create attribute and action functions
Call mkdir in the actions or attributes folder to create a "pending"
action or attribute function. A file "PENDING_BINDINGS_EDIT_ME.json"
will appear in the new folder. Edit that file with a correct action or
correct output location for a function, and the save will create the
pending function.

You can also configure function arguments in the bindings for attribute
functions now.