Move generated ApprovalRequirement for Schema category to individual Schemas
`ApprovalRequirement` are intended to have the `EntityId` of the thing
that has changed, and is requiring approval, not the category of the
thing that has changed. By having the `EntityId` of the thing that has
changed, we are able to provide information for where to look to see the
specific thing that has changed, instead of "there is an X that has
changed".