Check to see if Components have been moved to other Views when removing a View
When removing a View, we make sure that it isn't the only View that has
Geometry for any of the Components in it. We weren't taking into account
that a Component might be moved from only being in the View we're
removing to being in another View (or multiple other views).

We now track which Components got new Geometry for Views other than the
one being removed, and if we see that the View being removed had the
only Geometry for the Component, we see if it is in the list of
Components that are getting new Geomerty. If it is in that list, then we
allow the removal to proceed as we won't be orphaning the Component from
all Views.