Consolidate virtual approval requirement generation
Right now, if we need to generate a virtual approval requirement at all,
it will always be the same rule. Rather than repeating the generation of
identical rules across many match arms this removes duplication by
condensing everything other than generation for
ApprovalRequirementDefinition into the same match arm.
ApprovalRequirementDefinition is a bit special at the moment and will
likely be the last thing removed from this logic as we fill out more
complex defaults & fallback logic for the requirement definitions of
things in the graph.