Do not filter requirements by kind and other fixes
This commit ensures that the InsertApprovalModel no longer filters
requirements based on entity kind. Now, it shows a requirment per
entity and the labelling is expanded for this feature. This is NOT the
final state we want to be in due to showing potentially too many and
overwhelming results for the same approver groups and individuals, but
this is a step in the right direction.

This commit also ensures that the approval status data is up to date
on the requester's side when requesting approvals. This is due to a
potential race or missing edge case when the change set moves from an
"open" to a "needsapproval" state. A corresponding "TODO" has been
added to the problem area, but the current fix is sufficient in the
interim.

In addition to these changes, both the requirements and actions are now
scrollable. We now also show the email for a user (though truncation may
not work at this time) as well as right padding for the status text
(e.g. "Waiting..." which has now been capitalized to match the other
states).

One final change, we now always allow users to reject the changes, even
if all approvals have been met.

Co-authored-by: Wendy Bujalski <wendy@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>