fix(clover): add secret props in asset definition func
When we create the schema spec, for schemas that consume secrets, we must add a prop and socket with the correct configuration.  However, we provide abstraction over this in the asset definition function, so we must detect when secret props are being used, build a `SecretProp`, ensuring we don't add the corresponding `InputSocket` as the `SecretProp` handles adding both the prop and the socket.