Add max deliver config to pinga (default is 1)
This commit adds a configurable max delivery field to pinga, which
defaults to "1".

First, why does it default to "1"? The single naxum handler in pinga
server relies on the default AckLayer and its inner logic is infallible.
How is that the case? The job exeuction code is wrapped with an
infallible wrapper.

There are two feasible ways that the handler can fail, which causes a
NACK to be sent via the code in the default AckLayer... the JSON
extractor can fail (error 422 on deserialization) or the subject is not
a valid UTF8 string. The other extractors are fallible, but their
default implementations immediately return "Ok". Therefore, either the
request payload is a poison pill or the subject is invalid. In either
case, we should not retry the job and we should immediately send the
message to the Dead Letter Queue stream, which is already configured to
handle this.

As a result of this paradigm, there is no custom backoff or increased
number of max deliveries. So... why does this commit make it
configurable? Due errant test behavior when adding a custom backoff
and deserialization errors recently seen when pulling messages off the
"PINGA_JOBS" stream, we have not confirmed whether or not the system
relies on re-deliveries at the pinga-level or not. Therefore, if we
discover that re-delivery is required, the max delivery should be
increased accordingly with no custom backoff.

An example max delivery increase would be from 1 (default) to 60,000.
WTF? WHY? Well, we have observed in live production scenarios that we
can see around ~500 re-deliveries per second. Setting max deliveries to
60,000 allows for approximately 2 minutes worth of pinga trying to
execute a job. This math is inadequate at best, but the point is this:
if we find that we do require retries, then bump that number up.

In addition to these changes, the "initial DVU" step of the rebaser has
been removed. It was added in the quiescence work, but it may actually
be the cause of the poison pill messages. Working with @britmyerss
revealed that it is likely unnecessary and both testing and static
analysis have indicated as such.

This was tested with 201 components with the root component being a
Region as a configurationDownFrame. There were VPCs, Subnets, EC2
Instances, and all kinds of nesting. Flipping the region in the root
set off pinga into a frenzy and none of the errors were observed.

Co-authored-by: Fletcher Nichol <fletcher@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>