feat(si-fs): asset def types
This adds a new file to the asset definition directories: `index.d.ts`,
containing all the types for the asset builder. It inserts a
triple-slash directive at the top of the asset index.ts file that
references these types, so typescript language servers will pick up the
types automatically. This directive will be removed from the code when
the file is saved, before it is sent to the backend.