Initial work for frontend data flow rearchitecture
This outlines an initial structure for defining materialized views
(atoms) that are generated on every change.

These atoms are kept in a KV store (`frigg`), along side an "index" of
all of the atoms that exist for the current state of each change set.

The materialized view definition that generates each atom is able to
specify that changes to a specific `EntityKind` should cause the atom to
be (re)generated, and are also able to specify that it "references"
other materialized view definitions. This "references" information is
used to build a dependency graph to ensure that no materialized view
that references another will be built before all of the atoms for the
referenced materialized view are built.

If the frontend requests the index of atoms for a change set from SDF,
and the change set is valid, but SDF/frigg is unable to find the index
for it, SDF will build all materialized views so that it can provide the
front end with the resulting index of atoms.

This also creates a new websocket (bifrost) that is used for
broadcasting JSON Patch information about atom changes whenever
materialized views are built, and there is a pre-existing index for that
change set, in order to help minimize the amount of data going over the
wire on every change for the front end to become aware the change.

A new pinia store (heimdall) controls a web worker that has access to
wasm sqlite where we persist data. We have database tables and an update
algorithm that keeps a single unique reference to an Atom based on its
EntityKind, id, and, checksum. This prevents duplicative data. And we
store the Atom's relationship to change sets through the snapshot
address. We accept both JSON patch data and "cold start" (mjolnir)
events over the bifrost web socket.

Heimdall exposes the bifrost querying interface to tan stack, which
is used within individual components. When we write to an Atom in
sqlite, we also fire a callback to Heimdall to bust the tan stack
query cache, which re-fires the queryFn, and re-renders component data.

We prune data from sqlite as old data becomes obsolete from being
patched, as well as change sets that are applied, abandoned or closed.

Co-authored-by: John Obelenus <jobelenus@systeminit.com>
Co-authored-by: Jacob Helwig <jacob@systeminit.com>
Co-authored-by: Fletcher Nichol <fletcher@systeminit.com>