Move special case funcs to intrinsics (round two)
This commit moves the "si:resourcePayloadToValue" and
"si:normalizeToArray" funcs to the intrinsics API and package. This was
primarily motivated by the performance upside: we no longer run these
funcs in veritech, but in whoever invokes the func runner, which is
pinga. Now, we move from hundreds of milliseconds of execution time to
nearly instant execution of these two funcs.

The Javascript-based functions are re-written in Rust with as close of
a one-to-one logic mapping as possible with "null" and "empty" values
and live type evaluation in play.

As a result of these changes, two major things have happened:

1) We have expanded the asset intrinsics panel in the customization
   interface to allow you to use "si:normalizeToArray" (though you will
   have to regenerate your asset to install the new func for the first
   time, which the UI will instruct you to do).
2) We are allowing the old special case funcs and the new intrinsics to
   co-exist.

Since a chunk of the system relies on func names being unique,
achieving those two major "things" was... a total pain in the ass. As
you comb this commit, you'll notice all the edge cases. However, we've
found them all... famous last words, etc. etc.

What else has changed? This commit fixes the ordering of intrinsic
output options in the corresponding panel. They were ordered by name,
but now that we are showing the full path (recent change unrelated to
this commit), the ordering now relies on the path. The panel also has
a fancy new dropdown to explain what each intrinsic func does.

It's worth noting: we now hide the old special case funcs from the user
for _any_ asset. They'll be auto-migrated upon regeneration, but you'll
notice that only "si:normalizeToArray" is configurable. This is
intentional. We are treating the new "si:resourcePayloadToValue" as an
internal-only func that we setup for you as part of creating an asset
with a real world resource.

Another side effect of these changes is that that test exclusive
schemas no longer need to create their own special case funcs. Now, that
will get set up automatically. This indirectly fixed some partially
finished schemas, like the "Starfield" schema, which is why you now
see that the approvals SDF integration test now accounts for
auto-enqueued actions. Pretty sick.

Co-authored-by: Brit Myers <brit@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>