Guarantee single schema, variant and that domain is object prop
We have a lot of repeated checks for there being 1 schema, 1 variant and
that domain is an object prop. This ensures that at the type level.

It also makes tooltips for individual prop types (object, string, etc.)
a bunch more readable by pulling out each PropSpec kind individually
(typescript then does the work ahead of time so you can just look at the
props on the kind you are looking for)