feat(si-fs): support deno
Supports deno by configuring a workspace root deno.json and a
.vscode/settings.json with "deno: enable". In order to support deno we
prefetch every function so we can fill out the deno.json workspaces
list.

If new functions come in, or are created, *you will have to reload your
deno workspace to get them parsed by the deno language server*.

`node_modules` for deno are not yet supported. They can be! But a
passthrough will have to be implemented so that the node module tree can
be written to disk in another folder, and mirrored in the
si-filesystem system.