feat(si-fs): categorize schemas, add env_logger
List schemas in category folders. You can make a new category with
`mkdir`, but it won't be a real category until you make a new schema in
that category (with `mkdir` as well, in the category directory).

You can also make a new category by setting the category in attrs.json.

We only list the categories for the *default* schema variants of a
schema. So if you change the category of an existing schema on the
unlocked attributes, it won't appear in that category in the filesystem
until you merge that change set with HEAD. For new schemas, where the
unlocked is the default, the schema will be relisted in the new category
folder on the next uncached read.

You can now get log output by launching si-fs with RUST_LOG=debug (or
trace, info, etc.) in the environment.