build(nix): create `si-fs-standalone` Nix pkg for a Nix-free binary
This change introduces a new Nix derivation helper
(`standaloneBinaryDerivation`) which uses a `binDerivation`-built
package and prepares its main binary to be runnable without the full Nix
package closure.

Note that at the moment, this very much depends on the
dynamic libraries at play and whether or not the executing system has
the correct versions. At the moment, we're targetting the `si-fs` binary
which only depends on `libc` and `libgcc`. Future work will attempt to
make these binaries more flexible and portable where possible.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>