feat: allow for refining discovery operations
This updates the discovery function to allow for refining the discovered
resources by pre-configuring the component before you call discovery. It
does a relatively naive deep equality match. For example, if you
configure a VPC ID for a Subnet, then discover all subnets, it will
return only the VPC IDs for that subnet.

In addition, it will replicate all connections the component you ran
discovery from to the discovered assets - so they will get credentials,
regions, etc.

Note that this doesn't currently work well if the component uses a
parent to infer the connections, because that data is not available yet
in management functions.